third_party/blink/web_tests/platform/webview/external/wpt/pointerevents/pointerevent_after_target_appended_interleaved.tentative_touch-expected.txt

This is a testharness.js-based test.
[FAIL] mouse events from touch received before/after child attached at pointerdown
  assert_equals: events received expected "mouseover@parent,mouseenter@parent,mousedown@parent,(child-attached),mouseover@child,mouseenter@child,mouseup@child,click@parent,mousedown@child,mouseup@child,click@child,mouseout@child,mouseleave@child,mouseleave@parent" but got "(child-attached),mouseover@child,mouseenter@parent,mouseenter@child,mousedown@child,mouseup@child,mousedown@child,mouseup@child,mouseout@child,mouseleave@child,mouseleave@parent"
[FAIL] mouse events from touch received before/after child attached at pointerup
  assert_equals: events received expected "mouseover@parent,mouseenter@parent,mousedown@parent,mouseup@parent,(child-attached),click@parent,mouseover@child,mouseenter@child,mousedown@child,mouseup@child,click@child,mouseout@child,mouseleave@child,mouseleave@parent" but got "(child-attached),mouseover@child,mouseenter@parent,mouseenter@child,mousedown@child,mouseup@child,mousedown@child,mouseup@child,mouseout@child,mouseleave@child,mouseleave@parent"
[FAIL] mouse events from touch received before/after child moved at pointerdown
  assert_equals: events received expected "mouseover@child,mouseenter@parent,mouseenter@child,mousedown@child,(child-moved),mouseover@child,mouseenter@child,mouseup@child,click@child,mousedown@child,mouseup@child,click@child,mouseout@child,mouseleave@child,mouseleave@parent" but got "(child-moved),mouseover@child,mouseenter@parent,mouseenter@child,mousedown@child,mouseup@child,mousedown@child,mouseup@child,mouseout@child,mouseleave@child,mouseleave@parent"
[FAIL] mouse events from touch received before/after child moved at pointerup
  assert_equals: events received expected "mouseover@child,mouseenter@parent,mouseenter@child,mousedown@child,mouseup@child,(child-moved),click@child,mouseover@child,mouseenter@child,mousedown@child,mouseup@child,click@child,mouseout@child,mouseleave@child,mouseleave@parent" but got "(child-moved),mouseover@child,mouseenter@parent,mouseenter@child,mousedown@child,mouseup@child,mousedown@child,mouseup@child,mouseout@child,mouseleave@child,mouseleave@parent"
Harness: the test ran to completion.
